[#1] - Renamed namespaces to be Stoic\Input and reorganized the file structure.
